requirements.txt
# video ingestion
yt-dlp

# computer vision
opencv-python
paddlepaddle
paddleocr
ultralytics        # YOLOv8-Pose

# clustering & analysis
numpy
pandas
scipy
tslearn             # DTW

# database
psycopg2-binary

# utils
tqdm
python-dotenv